Resize headers on partial parse
The main benefit of this is that it enables you to get the parsed
headers from partial results when using uninitialized headers. Prior to
this, on a partial result, the headers in a request or response would be
replaced with the original slice passed when creating the request or
response (typically an empty slice when using uninitialized headers),
rather than the initialized portion of the uninitialized headers passed
into the parse function call.

It also makes handling partial parse results more convenient in general
since you can iterate over the headers in the request or response
without having to manually determine the last element.

This is technically a breaking change, though it seems unlikely that any
client would depend on the previous behavior in a way that the change
would break their code.